Swipe delete for each recycler and deleting expired todo list

+++ app/src/main/java/com/vanard/tulis/TodoActivity.java
@@ -17,6 +17,7 @@ import android.support.v7.widget.helper.ItemTouchHelper;
 import android.util.Log;
 import android.view.View;
 import android.widget.TextView;
+import android.widget.Toast;
 
 import com.google.android.gms.tasks.OnSuccessListener;
 import com.google.firebase.auth.FirebaseAuth;
@@ -60,6 +61,8 @@ public class TodoActivity extends AppCompatActivity implements RecyclerItemTouch
     private ArrayList<TodoLayout> todoTomorrow;
     private ArrayList<TodoLayout> todoOtherDay;
     private TodoRecyclerAdapter adapter;
+    private TodoRecyclerAdapter adapter1;
+    private TodoRecyclerAdapter adapter2;
 
     public String catDId;
     public String catTit;
@@ -135,40 +138,50 @@ public class TodoActivity extends AppCompatActivity implements RecyclerItemTouch
                 todoLayoutList.clear();
                 todoTomorrow.clear();
                 todoOtherDay.clear();
+
                 for (DocumentChange doc : queryDocumentSnapshots.getDocumentChanges()) {
                     if (doc.getType() == DocumentChange.Type.ADDED) {
                         TodoLayout todoLayout = doc.getDocument().toObject(TodoLayout.class);
 
+                        String docu = todoLayout.getDocumentId();
+                        String catDocu = todoLayout.getCategory_documentId();
+                        
                         Calendar c = Calendar.getInstance();
 
                         long deadline = todoLayout.getDate().getTime();
                         long deadtime = todoLayout.getDeadtime();
 
                         if (c.get(Calendar.DATE) == deadtime && (c.getTimeInMillis() - deadline) <= 0 && (c.getTimeInMillis() - deadline) > (-1*milDay)){
+                            Log.d(TAG, "onEvent: hari ini");
                             todoLayoutList.add(todoLayout);
                             if (todoLayoutList.size() > 0)
                                 today.setVisibility(View.VISIBLE);
 
                             adapter.notifyDataSetChanged();
-                            return;
                         }
-                        if (c.get(Calendar.DATE) - deadtime == -1 && (c.getTimeInMillis() - deadline) <= (-1*milDay)  && (c.getTimeInMillis() - deadline) >= (-2*milDay)){
+                        if (c.get(Calendar.DATE) - deadtime == -1 && (c.getTimeInMillis() - deadline) >= (-2*milDay)){
+                            Log.d(TAG, "onEvent: besok");
                             todoTomorrow.add(todoLayout);
                             if(todoTomorrow.size() > 0)
                                 tomorrow.setVisibility(View.VISIBLE);
-                            adapter.notifyDataSetChanged();
-                            return;
+                            adapter1.notifyDataSetChanged();
                         }
                         if (c.get(Calendar.DATE) - deadtime < -1 && (c.getTimeInMillis() - deadline) <= (-1*milDay)){
+                            Log.d(TAG, "onEvent: other");
                             todoOtherDay.add(todoLayout);
                             if (todoOtherDay.size() > 0)
                                 aday.setVisibility(View.VISIBLE);
-                            adapter.notifyDataSetChanged();
-                            return;
+                            adapter2.notifyDataSetChanged();
+                        }if (c.getTimeInMillis() - deadline > 0) {
+                            Log.d(TAG, "onEvent: terlewat");
+                            db.document("Category/" + catDocu + "/Todo/" + docu).delete().addOnSuccessListener(new OnSuccessListener<Void>() {
+                                @Override
+                                public void onSuccess(Void aVoid) {
+                                    Toast.makeText(TodoActivity.this, "Waktunya sudah lewat", Toast.LENGTH_SHORT).show();
+                                }
+                            });
                         }
-
 //                        todoLayoutList.add(todoLayout);
-
                     }
                 }
             }
@@ -189,19 +202,50 @@ public class TodoActivity extends AppCompatActivity implements RecyclerItemTouch
 
     @Override
     public void onSwiped(RecyclerView.ViewHolder viewHolder, int direction, int position) {
+        TodoRecyclerAdapter.ViewHolder a = (TodoRecyclerAdapter.ViewHolder) viewHolder;
+        int s = a.id;
         if (viewHolder instanceof TodoRecyclerAdapter.ViewHolder) {
-            // remove the item from recycler view
-            String doc = todoLayoutList.get(position).getDocumentId();
-            String catDoc = todoLayoutList.get(position).getCategory_documentId();
-            db.document("Category/" + catDoc + "/Todo/"+ doc).delete().addOnSuccessListener(new OnSuccessListener<Void>() {
-                @Override
-                public void onSuccess(Void aVoid) {
-                }
-            });
-            adapter.removeItem(viewHolder.getAdapterPosition());
-            adapter.notifyItemRemoved(position);
-            adapter.notifyDataSetChanged();
-            initData();
+            if (s == 1){
+                String doc = todoLayoutList.get(position).getDocumentId();
+                String catDoc = todoLayoutList.get(position).getCategory_documentId();
+                db.document("Category/" + catDoc + "/Todo/"+ doc).delete().addOnSuccessListener(new OnSuccessListener<Void>(){
+                    @Override
+                    public void onSuccess(Void aVoid) {
+                    }
+                });
+                adapter.removeItem(viewHolder.getAdapterPosition());
+                adapter.notifyItemRemoved(position);
+                adapter.notifyDataSetChanged();
+                initData();
+                Log.d(TAG, "onSwiped: today hapus");
+            }
+            if (s == 2){
+                String doc = todoTomorrow.get(position).getDocumentId();
+                String catDoc = todoTomorrow.get(position).getCategory_documentId();
+                db.document("Category/" + catDoc + "/Todo/"+ doc).delete().addOnSuccessListener(new OnSuccessListener<Void>() {
+                    @Override
+                    public void onSuccess(Void aVoid) {
+                    }
+                });
+                adapter1.removeItem(viewHolder.getAdapterPosition());
+                adapter1.notifyItemRemoved(position);
+                adapter1.notifyDataSetChanged();
+                initData();
+                Log.d(TAG, "onSwiped: tomorrow hapus");
+            }if (s == 3){
+                String doc = todoOtherDay.get(position).getDocumentId();
+                String catDoc = todoOtherDay.get(position).getCategory_documentId();
+                db.document("Category/" + catDoc + "/Todo/"+ doc).delete().addOnSuccessListener(new OnSuccessListener<Void>() {
+                    @Override
+                    public void onSuccess(Void aVoid) {
+                    }
+                });
+                adapter2.removeItem(viewHolder.getAdapterPosition());
+                adapter2.notifyItemRemoved(position);
+                adapter2.notifyDataSetChanged();
+                initData();
+                Log.d(TAG, "onSwiped: other day hapus");
+            }
         }
     }
 
@@ -219,7 +263,7 @@ public class TodoActivity extends AppCompatActivity implements RecyclerItemTouch
         rcv_today.setItemAnimator(new DefaultItemAnimator());
         rcv_today.addItemDecoration(new DividerItemDecoration(this, DividerItemDecoration.VERTICAL));
 
-        adapter = new TodoRecyclerAdapter(todoLayoutList, this);
+        adapter = new TodoRecyclerAdapter(todoLayoutList, this, 1);
         rcv_today.setAdapter(adapter);
 
         ItemTouchHelper.SimpleCallback itemTouchHelperCallback = new RecyclerItemTouchHelper(0, ItemTouchHelper.LEFT, this);
@@ -230,8 +274,8 @@ public class TodoActivity extends AppCompatActivity implements RecyclerItemTouch
         rcv_tomorrow.setItemAnimator(new DefaultItemAnimator());
         rcv_tomorrow.addItemDecoration(new DividerItemDecoration(this, DividerItemDecoration.VERTICAL));
 
-        adapter = new TodoRecyclerAdapter(todoTomorrow, this);
-        rcv_tomorrow.setAdapter(adapter);
+        adapter1 = new TodoRecyclerAdapter(todoTomorrow, this, 2);
+        rcv_tomorrow.setAdapter(adapter1);
 
         ItemTouchHelper.SimpleCallback itemTomorrow = new RecyclerItemTouchHelper(0, ItemTouchHelper.LEFT, this);
         new ItemTouchHelper(itemTomorrow).attachToRecyclerView(rcv_tomorrow);
@@ -241,8 +285,8 @@ public class TodoActivity extends AppCompatActivity implements RecyclerItemTouch
         rcv_other.setItemAnimator(new DefaultItemAnimator());
         rcv_other.addItemDecoration(new DividerItemDecoration(this, DividerItemDecoration.VERTICAL));
 
-        adapter = new TodoRecyclerAdapter(todoOtherDay, this);
-        rcv_other.setAdapter(adapter);
+        adapter2 = new TodoRecyclerAdapter(todoOtherDay, this, 3);
+        rcv_other.setAdapter(adapter2);
 
         ItemTouchHelper.SimpleCallback itemOther = new RecyclerItemTouchHelper(0, ItemTouchHelper.LEFT, this);
         new ItemTouchHelper(itemOther).attachToRecyclerView(rcv_other);
